requirements.txt
gevent
redis
Flask
requests
uwsgi
pytest
sqlalchemy
pg8000
pycrypto
